+++ CHANGELOG.md
@@ -48,3 +48,13 @@
 ## 0.4.2
 * Fix file picker for iOS and Android.
 * Keep Session.
+
+## 0.4.3
+* Make Drawer for account setting.
+* Fix bug unfollow button does not work.
+* App Password API
+
+## 0.4.4
+* Register account without invite code.
+* Password reset.
+* Delete account.